feat: Implement ML-5 Algorithm Performance Tracking + TODO cleanup
ML-5: Algorithm Performance Tracking

Adds execution time metrics for all spam detection algorithms to identify
performance bottlenecks and optimize slow checks.

Data Collection Layer:
- Change ProcessingTimeMs from long to double in ContentCheckResponseV2,
  ContentCheckResponse, and ContentCheckResult models (no DB migration needed,
  JSONB stores numbers generically, fractional precision matters for fast checks <10ms)
- Add Stopwatch.GetTimestamp() / GetElapsedTime() timing to all 11
  IContentCheckV2 implementations (73 return statements, inline pattern)
- Create GIN index on detection_results.check_results_json for fast JSONB queries
  (configured in AppDbContext, supports jsonb_array_elements performance)

Analytics Layer:
- Add AlgorithmPerformanceStats DTO model with metrics:
  TotalExecutions, AverageMs, P95Ms, MinMs, MaxMs, TotalTimeContribution
  PerformanceRating: Fast (<100ms), Medium (100-500ms), Slow (>500ms)
- Implement GetAlgorithmPerformanceStatsAsync in AnalyticsRepository
  Uses raw SQL with jsonb_array_elements() and PERCENTILE_CONT()
  No EF Core 10/Npgsql 10 equivalent for PostgreSQL-specific JSONB functions
  Leverages GIN index for performance
- Add Algorithm Performance Breakdown section to Analytics Performance tab
  Table showing per-algorithm execution times, sorted by total contribution
  Color-coded performance ratings (green/yellow/red chips)
  Formats: <1ms shows decimals, >=1000ms converts to seconds
  Parallel loading with existing analytics queries (5 tasks in Task.WhenAll)

Stopwatch Pattern:
All checks use inline pattern for consistency (no temporary variables):
ProcessingTimeMs = Stopwatch.GetElapsedTime(startTimestamp).TotalMilliseconds
Fixed 31 instances across 7 files that used var elapsed, elapsed2, etc.

Files Modified - Models (3):
- TelegramGroupsAdmin.ContentDetection/Models/ContentCheckResponseV2.cs
- TelegramGroupsAdmin.ContentDetection/Models/ContentCheckResponse.cs
- TelegramGroupsAdmin.ContentDetection/Models/ContentCheckResult.cs

Files Modified - Check Implementations (11):
- BayesSpamCheckV2, StopWordsSpamCheckV2, SimilaritySpamCheckV2
- CasSpamCheckV2, ThreatIntelSpamCheckV2, UrlBlocklistSpamCheckV2
- SpacingSpamCheckV2, InvisibleCharsSpamCheckV2
- OpenAIContentCheckV2, ImageSpamCheckV2, VideoSpamCheckV2, FileScanningCheckV2

Files Modified - Data Layer:
- TelegramGroupsAdmin.Data/AppDbContext.cs (GIN index configuration)
- Migration: 20251125181219_AddGinIndexOnCheckResultsJson

Files Modified - Analytics Layer:
- TelegramGroupsAdmin.Telegram/Models/AlgorithmPerformanceStats.cs (new)
- TelegramGroupsAdmin.Telegram/Repositories/IAnalyticsRepository.cs
- TelegramGroupsAdmin.Telegram/Repositories/AnalyticsRepository.cs

Files Modified - UI:
- TelegramGroupsAdmin/Components/Shared/Analytics/PerformanceMetrics.razor

TODO Cleanup:

Removed 4 obsolete TODO comments:
1. ContentDetectionService.cs - Fixed comment about ContentCheckRequest parameter
2. ModerationReportCard.razor - Removed note about markdown rendering (working)
3. StopWordRecommendationService.cs - Updated to reflect ProcessingTimeMs tracking
4. SeoPreviewScraper.cs - Removed completed URL normalization note

Created GitHub issues for 3 feature requests (#99, #100, #101).

+++ TelegramGroupsAdmin.ContentDetection/Models/ContentCheckResponse.cs
@@ -40,5 +40,5 @@ public record ContentCheckResponse
     /// <summary>
     /// Processing time for this check in milliseconds
     /// </summary>
-    public long ProcessingTimeMs { get; init; } = 0;
+    public double ProcessingTimeMs { get; init; } = 0;
 }
